# Optimize to_binary() function and to_bytes() method