@skirpichev 2024-12-26
N.B. this patch has two helpers, added to src/gmpy2_macros.h.  I'm not sure if it's a right place or they correctly named.